benchmark + C sequential

+++ Programming techniques/life/java/ILifeGenerator.java
@@ -6,5 +6,5 @@ public interface ILifeGenerator {
     /**
      * Computes the next state of the board using the original board.
      */
-    public short[][] compute(Life origin) throws Exception;
+    public byte[][] compute(Life origin) throws Exception;
 }
